requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.0
Pillow>=8.0.0
opencv-python>=4.4.0
tqdm>=4.50.0
PyQt5>=5.15.0
pyinstaller>=4.0 